Module correlation_track::ring_buffer

Ring buffer for transferring audio data between threads in a wait-free manner.

Functions as a single-producer single-consumer FIFO queue, where whole slices can be pushed and read at once.

All data is stored in one contiguous array, where read_index and write_index wrap around to the beginning if they reach the end. That is why it is called a “ring”.

Structs

Consumer

The Consumer party of a ring buffer. Can only pop data off the queue.

Producer

The Producer party of a ring buffer. Can only push data to the queue.

Traits

Item

Defines the interface that an item stored in the queue must implement. Only very simple types, such as floating point numbers or small structs that implement Copy are permitted.

Functions

with_capacity

Construct a ring buffer with a given capacity, and return (producer, consumer). The producer can then be sent to another thread for example.